Python: split test as suggested in review
